modify action creator join. and modify test.
